GRIBDatasets

Documentation for GRIBDatasets.

GRIBDatasets.GRIBDataset — Type

GRIBDataset{T, N}

Mapping of a GRIB file to a structure that follows the CF conventions.

It can be created with the path to the GRIB file:

ds = GRIBDataset(example_file);

The list of variables can be accessed:

julia> keys(ds)
7-element Vector{String}:
 "longitude"
 "latitude"
 "number"
 "valid_time"
 "level"
 "z"
 "t"

We can then index any of the variables or dimensions:

julia> z = ds["z"]
Variable `z` with dims:
Dimensions:
         longitude = 120
         latitude = 61
         number = 10
         valid_time = 4
         level = 2

julia> ds["number"] |> collect
10-element Vector{Int64}:
 0
 1
 2
 3
 4
 5
 6
 7
 8
 9

We can slice the variables along the dimensions:

julia> z[1:4, 3:6, 1, 1:2, 1]
4×4×2 reshape(::Array{Union{Missing, Float64}, 5}, 4, 4, 2) with eltype Union{Missing, Float64}:
[:, :, 1] =
 51038.7  50873.7  50731.2  50824.5
 51058.0  50859.0  50656.2  50682.0
 51077.5  50838.5  50578.5  50541.2
 51095.0  50807.2  50490.5  50398.5

[:, :, 2] =
 51031.0  50822.3  50672.5  50773.0
 51067.5  50820.5  50582.8  50647.5
 51102.0  50816.0  50471.3  50526.0
 51133.3  50806.3  50351.3  50399.3

GRIBDatasets.FileIndex — Type

Store for the messages of a GRIB file. Keeps track of the offset of the GRIB messages so they can be easily seeked. The unique_headers property gives all the different values for the keys in the GRIB file.

GRIBDatasets.MessageIndex — Type

MessageIndex

Stored information about a GRIB message. The keys can be accessed with getindex. The message offset and length are stored as property of the struct.