GRIBDatasets

Documentation for GRIBDatasets.

GRIBDatasets.COMPUTED_KEYS

GRIBDatasets.DEFAULT_EPOCH

GRIBDatasets.GRIB_STEP_UNITS_TO_SECONDS

GRIBDatasets.DiskValues

GRIBDatasets.DiskValues

GRIBDatasets.FileIndex

GRIBDatasets._add_headers!

GRIBDatasets.build_valid_time

GRIBDatasets.build_valid_time

GRIBDatasets.build_valid_time

GRIBDatasets.build_valid_time

GRIBDatasets.build_valid_time

GRIBDatasets.from_grib_date_time

GRIBDatasets.from_grib_date_time

GRIBDatasets.from_grib_date_time

GRIBDatasets.from_grib_month

GRIBDatasets.from_grib_step

GRIBDatasets.message_indices

GRIBDatasets.read_message

GRIBDatasets.COMPUTED_KEYS — Constant

Dictionary which maps a key to a conversion method. The first function is the 'to' conversion, the second is 'from'.

TODO: Actually applying the from_grib_step function results in different values to cfgrib.py, so step -> (from_grib_step, to_grib_step) is currently disabled.

Currently converts:

    "time" => (from_grib_date_time, to_grib_date_time)

    "valid_time" => (
        message -> from_grib_date_time(message, date_key="validityDate", time_key="validityTime"),
        message -> to_grib_date_time(message, date_key="validityDate", time_key="validityTime"),
    )

    "verifying_time" => (from_grib_month, m -> throw(ErrorException("Unimplemented")))

    "indexing_time" => (
        message -> from_grib_date_time(message, date_key="indexingDate", time_key="indexingTime"),
        message -> to_grib_date_time(message, date_key="indexingDate", time_key="indexingTime"),
    )

Example

A GRIB message containing 20160501 as the date key and 0 as the time key would end up calling:

julia> CfGRIB.COMPUTED_KEYS["time"](20160501, 0)
1462060800

GRIBDatasets.DEFAULT_EPOCH — Constant

Dates.DateTime

Default epoch used for from_ and to_ methods, set to 1970-01-01T00:00:00.

GRIBDatasets.GRIB_STEP_UNITS_TO_SECONDS — Constant

Vector{Union{Missing, Int64}}

Array used to convert the grib step units to seconds. As Julia is 1-indexed, not 0 like Python, you should take care to correctly access the array, typically just +1 to the step units before using it as an index.

Taken from eccodes stepUnits.table.

GRIBDatasets.DiskValues — Type

DiskValues{T, N, M} <: DA.AbstractDiskArray{T, N}

Object that maps the dimensions lookup to GRIB messages offsets. message_dims are the dimensions that are found in the GRIB message (namely longitudes and latitudes). other_dims are the dimensions that have been infered from reading the GRIB file index.

Example

julia> dv.other_dims
Dimensions:
         number = 10
         level = 2

julia> dv.offsets[3, 2]
324720

GRIBDatasets.DiskValues — Method

DiskValues(layer_index::FileIndex{T}, dims::Dimensions) where T

Create a DiskValues object from matching the GRIB messages headers in layer_index to the dimensions values in dims.

GRIBDatasets.FileIndex — Type

Store for indices of a GRIB file

GRIBDatasets._add_headers! — Method

Push the values of the message headers if they don't exist in the dictionnary d

GRIBDatasets.build_valid_time — Function

Returns a pair of (dims, data) based on the type of input

GRIBDatasets.build_valid_time — Method

julia> CfGRIB.build_valid_time(10, 10)
((), 36010)

GRIBDatasets.build_valid_time — Method

julia> CfGRIB.build_valid_time(1, [10])
(("step",), [36001])

GRIBDatasets.build_valid_time — Method

julia> CfGRIB.build_valid_time([10], 10)
(("time",), [36010])

GRIBDatasets.build_valid_time — Method

julia> CfGRIB.build_valid_time([10, 10], [10, 10])
(("time", "step"), [36010 36010; 36010 36010])

julia> CfGRIB.build_valid_time([10], [10])
((), 36010)

GRIBDatasets.from_grib_date_time — Function

Returns the integer seconds from epoch to the given date and time.

GRIBDatasets.from_grib_date_time — Method

Pulls out the date and time from given keys and passes them to from_grib_date_time(::Int, ::Int)

GRIBDatasets.from_grib_date_time — Method

GRIBDatasets.from_grib_month — Function

Returns the integer seconds from the epoch to the verifying month value in the GRIB message.

GRIBDatasets.from_grib_step — Function

Returns the step_key value in hours.

Uses GRIB_STEP_UNITS_TO_SECONDS to convert the step values to seconds, then divides by 3600.0 to get hours.

GRIBDatasets.message_indices — Method

message_indices(index::FileIndex, mind::MessageIndex, dims::Dimensions)

Find at which indices in dims correspond each GRIB message in index.

GRIBDatasets.read_message — Method

Reads a specific key from a GRIB.jl message. Attempts to convert the raw value associated with that key using the COMPUTED_KEYS mapping to from_grib_* functions.